+++ rubiks/src/model/pattern_database/thistlethwaite/g1.rs
@@ -37,7 +37,7 @@ impl PatternDatabase for G1PatternDatabase {
         todo!()
     }
 
-    fn set_num_moves_by_index(&mut self, ind: i32, num_moves: u8) -> bool {
+    fn set_num_moves_by_index(&mut self, ind: u8, num_moves: u8) -> bool {
         todo!()
     }
