+++ rubiks/src/model/pattern_database/thistlethwaite/g2.rs
@@ -65,7 +65,7 @@ impl PatternDatabase for G2PatternDatabase {
         todo!()
     }
 
-    fn set_num_moves_by_index(&mut self, ind: i32, num_moves: u8) -> bool {
+    fn set_num_moves_by_index(&mut self, ind: u8, num_moves: u8) -> bool {
         todo!()
     }
